Split generic template mutability tests
