+++ tests/docs_truth/repo_hygiene/file_size.rs
@@ -1,4 +1,5 @@
 mod focused_tests;
+mod function_method_template_splits;
 mod lexer_and_monomorphize;
 mod thresholds;
 mod typechecker_test_splits;
